+++ flexmeasures/api/v3_0/sensors.py
@@ -395,9 +395,9 @@ class SensorAPI(FlaskView):
         **kwargs,
     ):
         """
-        Trigger FlexMeasures to create a schedule.
+        Trigger FlexMeasures to create a schedule for a single flexible device, possibly taking into account inflexible devices.
 
-        .. :quickref: Schedule; Trigger scheduling job
+        .. :quickref: Schedule; Trigger scheduling job for one device
 
         Trigger FlexMeasures to create a schedule for this sensor.
         The assumption is that this sensor is the power sensor on a flexible asset.
@@ -411,9 +411,8 @@ class SensorAPI(FlaskView):
         For details on flexibility model and context, see :ref:`describing_flexibility`.
         Below, we'll also list some examples.
 
-        .. note:: This endpoint does not support to schedule an EMS with multiple flexible sensors at once. This will happen in another endpoint.
-                  See https://github.com/FlexMeasures/flexmeasures/issues/485. Until then, it is possible to call this endpoint for one flexible endpoint at a time
-                  (considering already scheduled sensors as inflexible).
+        .. note:: To schedule an EMS with multiple flexible sensors at once,
+                  use `this endpoint <../api/v3_0.html#post--api-v3_0-assets-(id)-schedules-trigger>`_ instead.
 
         The length of the schedule can be set explicitly through the 'duration' field.
         Otherwise, it is set by the config setting :ref:`planning_horizon_config`, which defaults to 48 hours.
@@ -556,8 +555,6 @@ class SensorAPI(FlaskView):
         except ValueError as err:
             return invalid_flex_config(str(err))
 
-        db.session.commit()
-
         response = dict(schedule=job.id)
         d, s = request_processed()
         return dict(**response, **d), s
@@ -579,7 +576,7 @@ class SensorAPI(FlaskView):
     ):
         """Get a schedule from FlexMeasures.
 
-        .. :quickref: Schedule; Download schedule from the platform
+        .. :quickref: Schedule; Download schedule for one device
 
         **Optional fields**
 
@@ -639,7 +636,6 @@ class SensorAPI(FlaskView):
                 )
                 return unrecognized_event(job.meta["fallback_job_id"], "fallback-job")
 
-        scheduler_info_msg = ""
         scheduler_info = job.meta.get("scheduler_info", dict(scheduler=""))
         scheduler_info_msg = f"{scheduler_info['scheduler']} was used."